requirements.txt
Django==3.2.8
djangorestframework==3.14.0
django-cors-headers==4.2.0
djangorestframework-simplejwt==5.3.1
django-filter==23.2